refactor: move logger
