+++ src/modules/logger.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { MyLogger } from 'src/services/logger.service';
+import { MyLogger } from 'src/loggers/logger';
 
 @Module({
   providers: [MyLogger],
